Introduce transaction execution listener to apply Oracle transaction priority.
