+++ data-r2dbc/src/main/java/io/micronaut/data/r2dbc/transaction/DefaultR2dbcReactorTransactionOperations.java
@@ -18,31 +18,27 @@ package io.micronaut.data.r2dbc.transaction;
 import io.micronaut.context.annotation.EachBean;
 import io.micronaut.context.annotation.Parameter;
 import io.micronaut.core.annotation.Internal;
+import io.micronaut.core.order.OrderUtil;
 import io.micronaut.data.connection.ConnectionStatus;
-import io.micronaut.data.connection.ConnectionSynchronization;
-import io.micronaut.data.connection.reactive.ReactiveConnectionStatus;
-import io.micronaut.data.connection.reactive.ReactiveConnectionSynchronization;
 import io.micronaut.data.r2dbc.connection.DefaultR2dbcReactorConnectionOperations;
 import io.micronaut.transaction.TransactionDefinition;
-import io.micronaut.transaction.annotation.OracleTransactional;
-import io.micronaut.transaction.exceptions.TransactionSystemException;
 import io.micronaut.transaction.reactive.ReactiveTransactionStatus;
 import io.micronaut.transaction.support.AbstractReactorTransactionOperations;
-import io.micronaut.transaction.support.TransactionUtil;
+import io.micronaut.transaction.support.ReactiveTransactionExecutionListener;
 import io.r2dbc.spi.Connection;
 import io.r2dbc.spi.ConnectionFactory;
 import io.r2dbc.spi.IsolationLevel;
-import io.r2dbc.spi.R2dbcException;
-import io.r2dbc.spi.Result;
+import jakarta.inject.Inject;
 import org.jspecify.annotations.Nullable;
 import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.time.Duration;
-import java.util.Locale;
 
 /**
  * Defines an implementation of Micronaut Data's core interfaces for R2DBC.
@@ -56,14 +52,22 @@ import java.util.Locale;
 final class DefaultR2dbcReactorTransactionOperations extends AbstractReactorTransactionOperations<Connection> implements R2dbcReactorTransactionOperations {
 
     private static final Logger LOG = LoggerFactory.getLogger(DefaultR2dbcReactorTransactionOperations.class);
-    private static final int ORACLE_INVALID_ALTER_SESSION_OPTION = 2248;
-    private static final String ORACLE_PRODUCT_NAME_UPPER = "ORACLE";
     private final String dataSourceName;
+    private final List<ReactiveTransactionExecutionListener<Connection>> transactionExecutionListeners;
+
+    @Inject
+    DefaultR2dbcReactorTransactionOperations(@Parameter String dataSourceName,
+                                             @Parameter DefaultR2dbcReactorConnectionOperations connectionOperations,
+                                             List<ReactiveTransactionExecutionListener<Connection>> transactionExecutionListeners) {
+        super(connectionOperations);
+        this.dataSourceName = dataSourceName;
+        this.transactionExecutionListeners = new ArrayList<>(transactionExecutionListeners);
+        OrderUtil.sort(this.transactionExecutionListeners);
+    }
 
     DefaultR2dbcReactorTransactionOperations(@Parameter String dataSourceName,
                                              @Parameter DefaultR2dbcReactorConnectionOperations connectionOperations) {
-        super(connectionOperations);
-        this.dataSourceName = dataSourceName;
+        this(dataSourceName, connectionOperations, Collections.emptyList());
     }
 
     @Override
@@ -89,17 +93,14 @@ final class DefaultR2dbcReactorTransactionOperations extends AbstractReactorTran
                 result = result.thenMany(connection.setTransactionIsolationLevel(isolationLevel));
             }
         }
-        OracleTransactional.Priority priority = TransactionUtil.getOraclePriority(definition);
-        if (priority != null && isOracleConnection(connection)) {
-            result = result.thenMany(applyOracleTxnPriority(connection, priority)
-                .flatMapMany(applied -> {
-                    if (applied) {
-                        registerOracleTxnPriorityReset(connectionStatus, connection);
-                    }
-                    return Flux.empty();
-                }));
+        for (ReactiveTransactionExecutionListener<Connection> transactionExecutionListener : transactionExecutionListeners) {
+            result = result.thenMany(transactionExecutionListener.beforeBegin(connectionStatus, definition));
         }
-        return result.thenMany(connection.beginTransaction());
+        result = result.thenMany(connection.beginTransaction());
+        for (ReactiveTransactionExecutionListener<Connection> transactionExecutionListener : transactionExecutionListeners) {
+            result = result.thenMany(transactionExecutionListener.afterBegin(connectionStatus, definition));
+        }
+        return result;
     }
 
     @Override
@@ -127,91 +128,6 @@ final class DefaultR2dbcReactorTransactionOperations extends AbstractReactorTran
             case SERIALIZABLE -> IsolationLevel.SERIALIZABLE;
             default -> null;
         }).orElse(null);
-    }
-
-    private boolean isOracleConnection(Connection connection) {
-        String productName = connection.getMetadata().getDatabaseProductName();
-        if (productName != null) {
-            productName = productName.toUpperCase(Locale.ENGLISH);
-        }
-        return ORACLE_PRODUCT_NAME_UPPER.equals(productName);
-    }
-
-    private Mono<Boolean> applyOracleTxnPriority(Connection connection, OracleTransactional.Priority level) {
-        return executeOracleTxnPriorityStatement(connection,
-            "ALTER SESSION SET \"txn_priority\"=\"" + level.name() + "\"",
-            "Setting",
-            level.name());
-    }
-
-    private Mono<Void> resetOracleTxnPriority(Connection connection) {
-        return executeOracleTxnPriorityStatement(connection,
-            "ALTER SESSION SET \"txn_priority\"=\"HIGH\"",
-            "Resetting",
-            "HIGH")
-            .onErrorMap(e -> new TransactionSystemException("Could not reset Oracle transaction priority", e))
-            .then();
-    }
-
-    private Mono<Boolean> executeOracleTxnPriorityStatement(Connection connection,
-                                                            String sql,
-                                                            String action,
-                                                            String level) {
-        if (LOG.isDebugEnabled()) {
-            LOG.debug("{} Oracle txn_priority to {}", action, level);
-        }
-        return Flux.from(connection.createStatement(sql).execute())
-            .flatMap(Result::getRowsUpdated)
-            .then(Mono.just(true))
-            .onErrorResume(e -> {
-                if (isOracleTxnPriorityUnsupported(e)) {
-                    LOG.debug("{} Oracle txn_priority failed with ORA-02248; continuing without priority support", action, e);
-                    return Mono.just(false);
-                }
-                return Mono.error(e);
-            });
-    }
-
-    private void registerOracleTxnPriorityReset(ConnectionStatus<Connection> connectionStatus, Connection connection) {
-        if (connectionStatus instanceof ReactiveConnectionStatus<Connection> reactiveConnectionStatus) {
-            reactiveConnectionStatus.registerReactiveSynchronization(new ReactiveConnectionSynchronization() {
-                @Override
-                public Publisher<Void> onComplete() {
-                    return resetOracleTxnPriority(connection);
-                }
-
-                @Override
-                public Publisher<Void> onError(Throwable throwable) {
-                    return resetOracleTxnPriority(connection);
-                }
-
-                @Override
-                public Publisher<Void> onCancel() {
-                    return resetOracleTxnPriority(connection);
-                }
-            });
-            return;
-        }
-        connectionStatus.registerSynchronization(new ConnectionSynchronization() {
-            @Override
-            public void executionComplete() {
-                resetOracleTxnPriority(connection).block();
-            }
-        });
-    }
-
-    private static boolean isOracleTxnPriorityUnsupported(Throwable exception) {
-        Throwable current = exception;
-        while (current != null) {
-            if (current instanceof R2dbcException r2dbcException) {
-                if (r2dbcException.getErrorCode() == ORACLE_INVALID_ALTER_SESSION_OPTION ||
-                    (r2dbcException.getMessage() != null && r2dbcException.getMessage().contains("ORA-02248"))) {
-                    return true;
-                }
-            }
-            current = current.getCause();
-        }
-        return false;
     }
 
     @Override
